+++ main.go
@@ -196,6 +196,15 @@ func main() {
 		os.Exit(1)
 	}
 
+	if err = (&controllers.CrossplaneClusterConfigReconciler{
+		Client:                mgr.GetClient(),
+		BaseDomain:            workloadClusterBaseDomain,
+		ManagementClusterName: managementClusterName,
+	}).SetupWithManager(ctx, mgr); err != nil {
+		setupLog.Error(err, "unable to create controller", "controller", "Frigate")
+		os.Exit(1)
+	}
+
 	if err := mgr.AddHealthzCheck("healthz", healthz.Ping); err != nil {
 		setupLog.Error(err, "unable to set up health check")
 		os.Exit(1)