+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add crossplane controller to create crossplane provider cluster config.
+
 ## [0.18.0] - 2025-06-16
 
 ### Added